Move files into new crates to create the new structure
