+++ crates/gears-app/Cargo.toml
@@ -1,17 +1,14 @@
-
 [package]
-name = "gears"
+name = "gears-app"
 version.workspace = true
-authors.workspace = true
 edition.workspace = true
+authors.workspace = true
 description = "The gears game engine library"
 homepage.workspace = true
 repository.workspace = true
 keywords.workspace = true
 license.workspace = true
 publish = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
 anyhow = "1.0"
