+++ crates/gears-ecs/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "gears-ecs"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
