+++ crates/gears-ecs/src/components/interactive.rs
