+++ crates/gears-ecs/src/components/lights.rs
