+++ crates/gears-macro/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "gears_macro"
-version = "0.1.0"
-edition = "2021"
+name = "gears-macro"
+version.workspace = true
+edition.workspace = true
 
 [lib]
 proc-macro = true
