+++ crates/gears-renderer/shaders/shader.wgsl
