+++ crates/gears-renderer/shaders/wireframe.wgsl
