+++ crates/gears-state/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "gears-state"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
