--- gears/src/core/mod.rs
@@ -1,6 +1,0 @@
-pub mod app;
-pub mod config;
-pub mod event;
-pub mod threadpool;
-
-pub type Dt = std::time::Duration;
